KAFKA-17238: Move VoterSet and ReplicaKey from raft.internals to raft
