+++ raft/src/main/java/org/apache/kafka/raft/ElectionState.java
@@ -17,7 +17,6 @@
 package org.apache.kafka.raft;
 
 import org.apache.kafka.raft.generated.QuorumStateData;
-import org.apache.kafka.raft.internals.ReplicaKey;
 
 import java.util.List;
 import java.util.Objects;
