+++ raft/src/main/java/org/apache/kafka/raft/VoterSet.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.raft.internals;
+package org.apache.kafka.raft;
 
 import org.apache.kafka.common.Node;
 import org.apache.kafka.common.Uuid;
@@ -22,7 +22,6 @@ import org.apache.kafka.common.feature.SupportedVersionRange;
 import org.apache.kafka.common.message.VotersRecord;
 import org.apache.kafka.common.network.ListenerName;
 import org.apache.kafka.common.utils.Utils;
-import org.apache.kafka.raft.Endpoints;
 
 import java.net.InetSocketAddress;
 import java.util.Collections;
@@ -49,7 +48,7 @@ import java.util.stream.Stream;
 public final class VoterSet {
     private final Map<Integer, VoterNode> voters;
 
-    VoterSet(Map<Integer, VoterNode> voters) {
+    public VoterSet(Map<Integer, VoterNode> voters) {
         if (voters.isEmpty()) {
             throw new IllegalArgumentException("Voters cannot be empty");
         }
