+++ raft/src/main/java/org/apache/kafka/raft/internals/RemoveVoterHandler.java
@@ -24,6 +24,8 @@ import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.raft.LeaderState;
 import org.apache.kafka.raft.LogOffsetMetadata;
 import org.apache.kafka.raft.RaftUtil;
+import org.apache.kafka.raft.ReplicaKey;
+import org.apache.kafka.raft.VoterSet;
 import org.apache.kafka.server.common.KRaftVersion;
 
 import org.slf4j.Logger;
