+++ raft/src/test/java/org/apache/kafka/raft/MockNetworkChannel.java
@@ -18,7 +18,6 @@ package org.apache.kafka.raft;
 
 import org.apache.kafka.common.network.ListenerName;
 import org.apache.kafka.common.protocol.ApiKeys;
-import org.apache.kafka.raft.internals.VoterSetTest;
 
 import java.util.ArrayList;
 import java.util.HashMap;
